requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
cssselect2==0.7.0
Deprecated==1.2.13
dj-static==0.0.6
Django==4.1.2
django-admin-interface==0.22.2
django-autoslug==1.9.8
django-ckeditor==6.5.1
django-colorfield==0.7.2
django-cors-headers==3.13.0
django-filer==2.2.3
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-js-asset==2.0.0
django-mptt==0.14.0
django-polymorphic==3.1.0
django-redis==5.2.0
djangorestframework==3.14.0
easy-thumbnails==2.8.3
essential-generators==1.0
exceptiongroup==1.0.4
gunicorn==20.1.0
iniconfig==1.1.1
lxml==4.9.1
packaging==21.3
pdfkit==1.0.0
Pillow==9.2.0
pluggy==1.0.0
psycopg2-binary==2.9.5
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.6
redis==4.3.4
reportlab==3.6.12
six==1.16.0
sqlparse==0.4.3
static3==0.7.0
svglib==1.4.1
tinycss2==1.2.1
tomli==2.0.1
Unidecode==1.1.2
webencodings==0.5.1
wrapt==1.14.1
